task

+++ task-19-8/task-19-8/Controllers/CategoriesController.cs
@@ -17,8 +17,9 @@ namespace task_19_8.Controllers
             _db = db;
         }
 
-
+        
         [HttpGet]
+        [Route("Get all categories")]
         public IActionResult Get()
         {
             var cart = _db.Categories.ToList();
@@ -27,36 +28,45 @@ namespace task_19_8.Controllers
         }
 
 
-        [HttpGet("{id}")]
-        public IActionResult Get(int id)
+        [HttpGet]   //route method  attribute
+        [Route(" Get one categories")]
+        public IActionResult Get( [FromQuery]  int id , [FromQuery] string name)
         {
-            var data = _db.Categories.Where(c => c.CategoryId == id).ToList();
+            var data = _db.Categories.Where(c => c.CategoryId == id && c.CategoryName == name).FirstOrDefault();
 
             return Ok(data);
         }
+
+
+        
 
 
         [HttpGet]
-        [Route("Category/OneCategory/{id:int:min(5)}")]
+        [Route("Category/OneCategory/{name}")]   // route attribute & parmiter
 
-        public IActionResult GetCategory(int id)
+        public IActionResult GetCategory( string name )
         {
-            var data = _db.Categories.Where(c => c.CategoryId == id).ToList();
+            var data = _db.Categories.Where(c => c.CategoryName == name ).FirstOrDefault();
 
             return Ok(data);
         }
 
 
-        [HttpDelete("{Id}")]
+        [HttpDelete]
+        [Route("delete one categories/{Id}")]
 
         public IActionResult delete(int Id)
         {
-            var y = _db.Categories.FirstOrDefault(c => c.CategoryId == Id);
+            var y = _db.Categories.Find(Id);
+            if (y != null)
+            {
 
-            _db.Categories.Remove(y);
-            _db.SaveChanges();
+                _db.Categories.Remove(y);
+                _db.SaveChanges();
 
-            return Ok(y);
+                return Ok(y);
+            }
+            return BadRequest();
         }
